+++ Scripts/GameData/IncrementalMinMaxInt.cs
@@ -4,7 +4,7 @@ public struct IncrementalMinMaxInt
     public MinMaxInt baseAmount;
     public MinMaxInt amountIncreaseEachLevel;
 
-    public MinMaxInt GetAmount(short level)
+    public MinMaxInt GetAmount(int level)
     {
         return baseAmount + (amountIncreaseEachLevel * (level - 1));
     }
